Add ELO system
Added a proper ELO system to calculate the amount of rating a player
could gain/lose after the match is concluded, it bases on both players\'
ratings to ensure faireness in rating changes.
This system works best for 1v1 games, but is adapted to work for any
game room size.