Fix elasticsearch crashing on Linux machines
Changed the config to make elasticsearch run in singe-node mode and
disabled memory locking, also set fixed memory size to the JVM, that was
enough to make the container run but is far from ideal.
Those settings shouldn't stay on prod environment, it's only a
workaround for the Linux machines in the school since we do not have
enough permissions to set memory limits.